Fixed a concurrency issue with FlippableDataContainer and updated code to Java 11
Updated project to Java 11 and removed dependencies to super-pom and commons to make it easier to include this project in custom projects. Reformated code to use default IntelliJ code formatting guidelines and updated a couple of tests to use Junit 5 instead of 4.
Fixed also a bug in the FlippableDataContainer, which was now renamed to FlippingDataContainer to be more in line with the English language, that could lose some entries as the value added to the queue was not atomic. This approach, however, is 10+ times slower (meassured by an updated test)