Update pom.xml
Replaced local snapshot version of maven-javadoc-plugin with the most recent one. This was only used to check whether the plugin had an issue or the Javadoc binary shipped with Java. It turned out that this is some bug with multi-release modules and the Javadoc binary itself (Bug-ID: 9060106; if it gets acknowledged by Oracle)